flujo de tarea avance

+++ src/app/home/flujo-de-tareas/ver-flujo-de-tareas/ver-flujo-de-tareas.component.ts
@@ -3,23 +3,27 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
 import { Flujo_tarea } from 'src/app/models/flujo_tareas';
 import { Usuario, UsuarioService } from 'src/app/services/usuario.service';
-import {
-  AgregarFlujo,
-  Flujo,
-  FlujoService,
-} from 'src/app/services/flujo.service';
-import {
-  TareaSub,
-  TareaSubordinadaService,
-} from 'src/app/services/tarea-subordinada.service';
+import { AgregarFlujo, Flujo, FlujoService } from 'src/app/services/flujo.service';
+import { AgregarTareaSub, EjecutarTareaSub, TareaSub, TareaSubordinadaService } from 'src/app/services/tarea-subordinada.service';
+import { EjecFlujoTarea, EjecFlujoTareaService } from 'src/app/services/ejec-flujo-tarea.service';
+import { AsigTareaService } from 'src/app/services/asig-tarea.service';
 
 @Component({
   selector: 'app-ver-flujo-de-tareas',
   templateUrl: './ver-flujo-de-tareas.component.html',
   styleUrls: ['./ver-flujo-de-tareas.component.css'],
-  providers: [FlujoService, UsuarioService, TareaSubordinadaService],
+  providers: [FlujoService, UsuarioService, TareaSubordinadaService,AsigTareaService, EjecFlujoTareaService],
 })
 export class VerFlujoDeTareasComponent implements OnInit {
+  mostrar_add_responsable: boolean = false;
+  id_entrada: string = "";
+  ejecFlujoTarea: EjecFlujoTarea[] = [];
+  TareaSub_formulario: FormGroup;
+  id_usuario_crear_ts: number | undefined;
+  nombre_usuario_crear_ts: string | undefined;
+  warning: boolean = false;
+
+
   ListarUsuario: Usuario[] = [];
   filtroUsuario = '';
   ListarFlujo: Flujo_tarea[] = [];
@@ -43,17 +47,24 @@ export class VerFlujoDeTareasComponent implements OnInit {
     private FlujoService: FlujoService,
     private UsuarioService: UsuarioService,
     private activeRouter: ActivatedRoute,
+    private EjecFlujoTareaService: EjecFlujoTareaService,
+    private AsigTareaService: AsigTareaService,
     private TareaSubordinadaService: TareaSubordinadaService
-  ) {}
+  ) {
+    this.TareaSub_formulario = this.fb.group({
+      nombre_ts: ['', Validators.required],
+      descripcion_ts: ['', Validators.required],
+      fecha_entrega_ts: ['', Validators.required],
+    });
+  }
 
   ngOnInit(): void {
-    this.listarUsuario();
-    this.listarTareaSub();
+    this.id_entrada = this.activeRouter.snapshot.params['id'];
+    
 
-    const id_entrada = this.activeRouter.snapshot.params['id'];
 
-    if (id_entrada) {
-      this.FlujoService.getFlujo(id_entrada).subscribe({
+    if (this.id_entrada) {
+      this.FlujoService.getFlujo(this.id_entrada).subscribe({
         next: (res: any) => {
           this.flujo = <any>res[0];
           console.log(res);
@@ -61,16 +72,59 @@ export class VerFlujoDeTareasComponent implements OnInit {
         error: (err) => console.log(err),
       });
     }
+    this.listarUsuario();
+    this.listarTareaSub();
+    this.listarAsigTarea();
+  }
+
+  listarAsigTarea() {
+    this.EjecFlujoTareaService.getVistaEjecFlujoTarea(this.id_entrada).subscribe(
+      (res) => {
+          this.ejecFlujoTarea = <any>res;
+          this.btn_add_responsable();
+      },
+    );
+  }
+  
+  btn_add_responsable(){
+    if(this.ejecFlujoTarea.length == 0)
+    {
+      this.mostrar_add_responsable = true;
+    }
+  }
+
+  crear_tarea_sub() {
+    const TAREASUB: EjecutarTareaSub = {
+      nombre_ts: this.TareaSub_formulario.get('nombre_ts')?.value,
+      descripcion_ts: this.TareaSub_formulario.get('descripcion_ts')?.value,
+      fecha_entrega_ts: this.TareaSub_formulario.get('fecha_entrega_ts')?.value,
+      porcentaje_avance_ts: 0,
+      estado_ts: "En curso",
+      id_ft: this.id_entrada,
+    };
+
+    const año = TAREASUB.fecha_entrega_ts?.substring(0,4);
+    const mes = TAREASUB.fecha_entrega_ts?.substring(5,7);
+    const dia = TAREASUB.fecha_entrega_ts?.substring(8,10);
+    TAREASUB.fecha_entrega_ts = (dia +"-"+mes+"-"+ año)
+
+      this.TareaSubordinadaService.addEjecucionTareaSub(TAREASUB).subscribe();
+      window.location.reload();
+  }
+
+  limpiarFormularioTareaSub()
+  {
+    this.TareaSub_formulario.reset();
+    this.id_usuario_crear_ts = undefined;
+    this.nombre_usuario_crear_ts = undefined;
   }
 
   listarTareaSub() {
-    this.TareaSubordinadaService.getTareasSub().subscribe(
+    this.TareaSubordinadaService.getTareaSubPorFT(this.id_entrada).subscribe(
       (res) => {
-        console.log(res);
-        this.ListarTareaSub = <any>res;
-      },
-      (err) => console.log(err)
-    );
+          this.ListarTareaSub = <any>res;
+        },
+      );
   }
 
   listarUsuario() {
@@ -83,10 +137,31 @@ export class VerFlujoDeTareasComponent implements OnInit {
     );
   }
   eliminar() {
-    this.FlujoService.deleteFlujo(<any>this.flujo.id_ft).subscribe(
+    if(this.ListarTareaSub.length <1){
+      this.FlujoService.deleteFlujo(<any>this.flujo.id_ft).subscribe(
+        (res) => {
+          this.router.navigate(['/flujo_de_tareas']);
+        },
+        );
+      }else{
+      console.log("Mensaje de error")
+      this.warning = true;
+
+    }
+  }
+  eliminar_responsable() {
+    this.AsigTareaService.deleteAsigTarea(<any>this.ejecFlujoTarea[0].id_ft).subscribe(
       (res) => {
-        console.log('flujo eliminado');
-        this.router.navigate(['/flujo_de_tareas']);
+        console.log('Responsable eliminado');
+      },
+      (err) => console.log(err)
+    );
+  }
+
+  eliminar_tarea_sub() {
+    this.TareaSubordinadaService.deleteTareasub(<any>this.ListarTareaSub[0].id_ts).subscribe(
+      (res) => {
+        console.log('Tarea subordinada eliminado');
       },
       (err) => console.log(err)
     );
